% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimadorRazon.R
\name{forestMAS}
\alias{forestMAS}
\title{Muestreo aleatorio simple (MAS) para la estimación de inventario forestales}
\usage{
forestMAS(Sitios, VarInt, tamSit, Superficie, ErrorDeseado)
}
\arguments{
\item{Sitios}{Etiqueta de sitios muestreados}

\item{VarInt}{Variable de interés para su estimación}

\item{tamSit}{Tamaño del sitio en m2}

\item{Superficie}{Area de estudio en ha}

\item{ErrorDeseado}{Error de muestreo deseado (para determinar el tamaño de muestra) en porcentaje}
}
\value{
Estimación del inventario mediante muestreo aleatorio simple
}
\description{
Muestreo aleatorio simple (MAS) para la estimación de inventario forestales
}
\examples{
#Datos de ejemplo
Sitios <- c(1, 1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4)
VarInt <- c(0.0271,  0.0576,  0.0376,  0.0505,  3.4839,  0.0231,  0.0769,  0.045,  0.0147,  0.0136,  0.0114)
tamSit <- 1000
Superficie <- 400
ErrorDeseado <- 0.1

}
