% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimadorRazon.R
\name{sum_na}
\alias{sum_na}
\title{Funci贸n de suma con omisi贸n de NA}
\usage{
sum_na(x)
}
\arguments{
\item{x}{Vector o lista para suma con omision de NA}
}
\value{
El resultado de la suma con omision de NA
}
\description{
Funci贸n de suma con omisi贸n de NA
}
\examples{
sum_na(x = c(34, 20, NA, 2))
}
